Revert "Rename solana-runtime to sealevel (#6239)"
This reverts commit 2e921437cda72bf742b912a69e6e32043d12addc.
