+++ runtime/src/lib.rs
@@ -1,5 +1,39 @@
-pub use sealevel::{
-    accounts, accounts_db, accounts_index, append_vec, bank, bank_client, bloom, epoch_schedule,
-    genesis_utils, loader_utils, message_processor, rent_collector, stakes, status_cache,
-    storage_utils, transaction_batch, transaction_utils,
-};
+pub mod accounts;
+pub mod accounts_db;
+pub mod accounts_index;
+pub mod append_vec;
+pub mod bank;
+pub mod bank_client;
+mod blockhash_queue;
+pub mod bloom;
+pub mod epoch_schedule;
+pub mod genesis_utils;
+pub mod loader_utils;
+pub mod message_processor;
+mod native_loader;
+pub mod rent_collector;
+mod serde_utils;
+pub mod stakes;
+pub mod status_cache;
+pub mod storage_utils;
+mod system_instruction_processor;
+pub mod transaction_batch;
+pub mod transaction_utils;
+
+#[macro_use]
+extern crate solana_metrics;
+
+#[macro_use]
+extern crate solana_vote_program;
+
+#[macro_use]
+extern crate solana_stake_program;
+
+#[macro_use]
+extern crate solana_bpf_loader_program;
+
+#[macro_use]
+extern crate serde_derive;
+
+extern crate fs_extra;
+extern crate tempfile;
