--- sealevel/Cargo.toml
@@ -1,48 +1,0 @@
-[package]
-name = "sealevel"
-version = "0.20.0"
-description = "Sealevel, a deterministic, parallel runtime"
-authors = ["Solana Maintainers <maintainers@solana.com>"]
-repository = "https://github.com/solana-labs/solana"
-license = "Apache-2.0"
-homepage = "https://solana.com/"
-edition = "2018"
-
-[dependencies]
-bincode = "1.2.0"
-bv = { version = "0.11.0", features = ["serde"] }
-byteorder = "1.3.2"
-fnv = "1.0.6"
-fs_extra = "1.1.0"
-lazy_static = "1.4.0"
-libc = "0.2.62"
-libloading = "0.5.2"
-log = "0.4.8"
-memmap = "0.6.2"
-rand = "0.6.5"
-rayon = "1.2.0"
-serde = { version = "1.0.101", features = ["rc"] }
-serde_derive = "1.0.101"
-serde_json = "1.0.41"
-solana-logger = { path = "../logger", version = "0.20.0" }
-solana-measure = { path = "../measure", version = "0.20.0" }
-solana-metrics = { path = "../metrics", version = "0.20.0" }
-solana-bpf-loader-api = { path = "../programs/bpf_loader_api", version = "0.20.0" }
-solana-bpf-loader-program = { path = "../programs/bpf_loader_program", version = "0.20.0" }
-solana-sdk = { path = "../sdk", version = "0.20.0" }
-solana-stake-api = { path = "../programs/stake_api", version = "0.20.0" }
-solana-stake-program = { path = "../programs/stake_program", version = "0.20.0" }
-solana-storage-api = { path = "../programs/storage_api", version = "0.20.0" }
-solana-vote-api = { path = "../programs/vote_api", version = "0.20.0" }
-solana-vote-program = { path = "../programs/vote_program", version = "0.20.0" }
-sys-info = "0.5.8"
-tempfile = "3.1.0"
-solana-rayon-threadlimit = { path = "../rayon-threadlimit", version = "0.20.0" }
-itertools = "0.8.0"
-
-[lib]
-crate-type = ["lib"]
-name = "sealevel"
-
-[dev-dependencies]
-solana-noop-program = { path = "../programs/noop_program", version = "0.20.0" }
